direct.evidence.plot {dmetar}	R Documentation

Plot for direct evidence proportions in a network meta-analysis using netmeta

Description

This function plots relevant measures quantifying the direct evidence proportion, mean path length and aggregated minimal parallelism of a frequentist network meta-analysis model generated by netmeta.

Usage

direct.evidence.plot(x, random=FALSE, comparison.label.size=2,
numeric.label.size=3, subplot.ratio=c(5, 1.3, 1.3))

Arguments

x
An object of class netmeta containing the results of a network meta-analysis using the netmeta function.

random
Logical. If set to TRUE, results for the random-effects model are displayed. If set to FALSE, results for the fixed-effect model are displayed. FALSE by default.

comparison.label.size
A numeric value for the size of comparison labels to be used in the plot. Default is 2.

numeric.label.size
A numeric value for the label size of numeric values to be used in the plot. Default is 3.

subplot.ratio
A numeric vector containing three numbers. Defines the width for each of the three subplots included in the plot (from left to right). Default is c(5,1.3,1.3).

Details

The function generates a plot containing three subplots displaying relevant characteristics to evaluate the reliability of effect size estimates within a network meta-analysis model.

Direct Evidence Proportion. This bar chart displays the proportion of direct evidence (orange) contained in each network estimate. It is of note that both direct and indirect evidence may contribute to the violation of the assumption of consistency underlying network meta-analysis models. Nevertheless, this plot allows to distinguish comparison estimates for which direct evidence was used, and to what extent, and comparisons which had to be inferred by indirect evidence alone.

Minimal Parallelism. This bar chart displays the minimum number of independent paths contributing to the effect estimate on an aggregated level. Large values of parallelism can be interpreted as supporting the robustness of the estimate.

Mean Path Length. This bar chart displays the mean path length, which characterizes the degree of indirectness of an estimate. Higher mean path lengths indicate less reliable estimates, given that more similarity assumptions have to be made when serially combining direct comparisons. Following König, Krahn and Binder (2013), comparisons with mean path lengths greater than two should be interpreted with caution. This threshold is displayed as a blue vertical line in the plot.

Value

data: A data.frame containing columns for the proportion of direct and indirect evidence of each comparison (proportion.direct and proportion.indirect), the mean path length (meanpath) and the minimal parallelism (minpar) for each comparison.

plot: The generated plot (if the function output was saved to an object).

Author(s)

Mathias Harrer & David Daniel Ebert

References

Harrer, M., Cuijpers, P., Furukawa, T.A, & Ebert, D. D. (2019). Doing Meta-Analysis in R: A Hands-on Guide. DOI: 10.5281/zenodo.2551803. Chapter 11.1

König J., Krahn U., Binder H. (2013): Visualizing the flow of evidence in network meta-analysis and characterizing mixed treatment comparisons. Statistics in Medicine, 32, 5414–29

See Also

netmeta, netmeasures

Examples

## Not run: 
# Load Senn2013 data from netmeta
suppressPackageStartupMessages(library(netmeta))
data(Senn2013)

# Conduct network meta-analysis (fixed-effects model)
nma = netmeta(TE, seTE, treat1, treat2, studlab,
              data=Senn2013, sm='MD', comb.random=FALSE)

# Generate the plot
dep = direct.evidence.plot(nma, random=FALSE, comparison.label.size = 1,
numeric.label.size=1, subplot.ratio=c(3,1,1))
dep
## End(Not run)

[Package dmetar version 0.0.9000 Index]