dmetar {dmetar}	R Documentation

dmetar: Companion R package for the guide 'Doing Meta-Analysis in R'

Description

dmetar serves as the companion R package for the guide Doing Meta-Analysis in R by Mathias Harrer, Pim Cuijpers, Toshi Furukawa and David Daniel Ebert.

Details

The package contains complementary functions to facilitate the conduction of meta-analyses using the meta, metafor, netmeta and gemtc packages.

Author(s)

Mathias Harrer mathias.harrer@fau.de, David Daniel Ebert david.daniel.ebert@gmail.com

See Also

Useful links:

https://dmetar.protectlab.org

[Package dmetar version 0.0.9000 Index]