eggers.test {dmetar}	R Documentation

Perform Egger's test of the intercept

Description

This function performs Egger's test of the intercept for funnel plot asymmetry using an object of class meta.

Usage

eggers.test(x)

Arguments

x
An object of class meta, generated by the metabin, metagen, metacont, metacor, metainc, or metaprop function.

Details

Performs Egger's test (Egger et al., 1997) for funnel plot asymmetry. The metabias function is called internally. Egger's test may lack the statistical power to detect bias when the number of studies is small. Sterne et al. (2011) recommend to perform funnel plot asymmetry tests only when k ≥q 10. A warning is therefore printed when the number of studies in the meta object is k < 10.

Value

Returns a list containing the following elements:

intercept: The intercept (bias).

llci: The lower bound of the 95% intercept confidence interval.

ulci: The upper bound of the 95% intercept confidence interval.

t: The t-statistic for the intercept test.

p: The p-value for Egger's test.

meta.obj: The meta-analysis object of class meta originally provided to the function.

Author(s)

Mathias Harrer & David Daniel Ebert

References

Harrer, M., Cuijpers, P., Furukawa, T.A, & Ebert, D. D. (2019). Doing Meta-Analysis in R: A Hands-on Guide. DOI: 10.5281/zenodo.2551803. Chapter 9.1

Egger M, Smith GD, Schneider M & Minder C (1997), Bias in meta-analysis detected by a simple, graphical test. BMJ, 315, 629–634.

Sterne, JAC et al. (2011), Recommendations for Examining and Interpreting Funnel Plot Asymmetry in Meta-Analyses of Randomised Controlled Trials. BMJ 343, 1, doi: 10.1136/bmj.d4002 .

See Also

metabias

Examples

# Create meta-analysis results using the 'metagen' function
suppressPackageStartupMessages(library(meta))
data(ThirdWave)
m = metagen(TE, seTE, studlab = paste(Author),
    data = ThirdWave, comb.random = FALSE, hakn=TRUE)

# Plug result into 'eggers.test' function
res.et <- eggers.test(m)

# Inspect the results
summary(res.et)

# Generate a funnel plot. This calls the 'funnel' function
# in 'meta' internally; additional parameters of this function can also
# be provided (see '?meta::funnel').
plot(res.et, bg = "lightblue")

[Package dmetar version 0.0.9000 Index]