Print, summary and plot methods for objects created using 'dmetar' functions

Print, summary and plot S3 methods for objects of class direct.evidence.plot, find.outliers, influence.analysis, multimodel.inference, pcurve, power.analysis, subgroup.analysis.mixed.effects, and sucra.

# S3 methods for class 'direct.evidence.plot'
summary(x)
print(x)
plot(x, ...)

# S3 methods for class 'find.outliers'
summary(x)
print(x)
plot(x, ...)
forest(x, ...)

# S3 methods for class 'influence.analysis'
summary(x)
print(x)
plot(x, which = "all")

# S3 methods for class 'multimodel.inference'
summary(x)
print(x)
plot(x)

# S3 methods for class 'pcurve'
summary(x)
print(x)

# S3 methods for class 'power.analysis'
summary(x)
print(x)
plot(x)

# S3 methods for class 'power.analysis.subgroup'
summary(x)
print(x)
plot(x)

# S3 methods for class 'subgroup.analysis.mixed.effects'
summary(x)
print(x)
plot(x, ...)
forest(x, ...)

# S3 methods for class 'sucra'
print(x)
plot(x)

# S3 method for direct.evidence.plot
summary(x)

# S3 method for find.outliers
print(x)

# S3 method for find.outliers
plot(x, ...)

# S3 method for InfluenceAnalysis
print(x)

# S3 method for InfluenceAnalysis
plot(x, which = "all")

# S3 method for multimodel.inference
print(x)

# S3 method for multimodel.inference
plot(x)

# S3 method for pcurve
print(x)

# S3 method for power.analysis
print(x)

# S3 method for power.analysis
plot(x)

# S3 method for power.analysis.subgroup
print(x)

# S3 method for power.analysis.subgroup
plot(x)

# S3 method for subgroup.analysis.mixed.effects
print(x)

# S3 method for subgroup.analysis.mixed.effects
plot(x, ...)

# S3 method for sucra
plot(x)

# S3 method for sucra
print(x)

Arguments

x
An object of class direct.evidence.plot, find.outliers, influence.analysis, multimodel.inference, pcurve, power.analysis, subgroup.analysis.mixed.effects, or sucra.

...
Other arguments of the meta.forest or metafor's codeforest function. Can be used for styling the generated forest plot.

Details

A total of four package-specific S3 methods are provided in dmetar: S3 methods for print, summary, plot and forest. Outputs from print and summary are always identical. When both plot and forest are available for a class, outputs of both functions are identical (i.e., a forest plot is returned). The forest/plot function allows additional arguments of the meta.forest or metafor's codeforest function (depending on the class of the meta-analysis object on which prior calculations are based on). These can be used for further styling of the forest plot.