summary.gosh.diagnostics {dmetar}	R Documentation

Print, summary and plot methods for objects created using 'dmetar' functions

Description

Print, summary and plot S3 methods for objects of class direct.evidence.plot, find.outliers, influence.analysis, multimodel.inference, pcurve, power.analysis, subgroup.analysis.mixed.effects, direct.evidence.plot, gosh.diagnostics and sucra.

Usage

## S3 method for class 'gosh.diagnostics'
summary(object, ...)

Arguments

object
An object of class gosh.diagnostics.

...
Additional arguments.

Details

A total of four package-specific S3 methods are provided in dmetar: S3 methods for print, summary, plot and forest. Outputs from print and summary are always identical. When both plot and forest are available for a class, outputs of both functions are identical (i.e., a forest plot is returned). The forest/plot function allows additional arguments of the meta.forest or metafor's codeforest function (depending on the class of the meta-analysis object on which prior calculations are based on). These can be used for further styling of the forest plot.

Author(s)

Mathias Harrer & David Daniel Ebert

[Package dmetar version 0.0.9000 Index]